refactor pipeline and models isp
