+++ models/br_rj_isp_estatisticas_seguranca/br_rj_isp_estatisticas_seguranca__armas_fogo_apreendidas_mensal.sql
@@ -4,7 +4,6 @@
         schema="br_rj_isp_estatisticas_seguranca",
     )
 }}
-
 select
     safe_cast(ano as int64) ano,
     safe_cast(mes as int64) mes,
