+++ models/br_rj_isp_estatisticas_seguranca/br_rj_isp_estatisticas_seguranca__evolucao_mensal_cisp.sql
@@ -1,5 +1,4 @@
 {{ config(alias="evolucao_mensal_cisp", schema="br_rj_isp_estatisticas_seguranca") }}
-
 select
     safe_cast(ano as int64) ano,
     safe_cast(mes as int64) mes,
