+++ models/br_rj_isp_estatisticas_seguranca/br_rj_isp_estatisticas_seguranca__evolucao_mensal_municipio.sql
@@ -3,7 +3,6 @@
         alias="evolucao_mensal_municipio", schema="br_rj_isp_estatisticas_seguranca"
     )
 }}
-
 select
     safe_cast(ano as int64) ano,
     safe_cast(mes as int64) mes,
