+++ pipelines/utils/crawler_isp/flows.py
@@ -1,0 +1,117 @@
+# -*- coding: utf-8 -*-
+"""
+Flows for br_rj_isp_estatisticas_seguranca.
+"""
+
+from datetime import timedelta
+
+from prefect import Parameter, case
+from prefect.run_configs import KubernetesRun
+from prefect.storage import GCS
+from prefect.tasks.prefect import create_flow_run, wait_for_flow_run
+
+from pipelines.constants import constants
+from pipelines.utils.constants import constants as utils_constants
+from pipelines.utils.crawler_isp.tasks import (
+    clean_data,
+    get_count_lines,
+)
+from pipelines.utils.decorators import Flow
+from pipelines.utils.execute_dbt_model.constants import (
+    constants as dump_db_constants,
+)
+from pipelines.utils.metadata.tasks import update_django_metadata
+from pipelines.utils.tasks import (
+    create_table_and_upload_to_gcs,
+    get_current_flow_labels,
+    rename_current_flow_run_dataset_table,
+)
+
+# ! Evolucao_mensal_cisp
+with Flow(
+    name="BD Template - Estatísticas de Segurança",
+    code_owners=[
+        "trick",
+    ],
+) as flow_isp:
+    dataset_id = Parameter(
+        "dataset_id", default="br_rj_isp_estatisticas_seguranca", required=True
+    )
+    table_id = Parameter("table_id", required=True)
+
+    # Materialization mode
+    target = Parameter("target", default="prod", required=False)
+    materialize_after_dump = Parameter(
+        "materialize_after_dump", default=True, required=False
+    )
+
+    dbt_alias = Parameter("dbt_alias", default=True, required=False)
+
+    rename_flow_run = rename_current_flow_run_dataset_table(
+        prefix="Dump: ",
+        dataset_id=dataset_id,
+        table_id=table_id,
+        wait=table_id,
+    )
+    update_metadata = Parameter(
+        "update_metadata", default=True, required=False
+    )
+
+    compararison_between_lines = get_count_lines(file_name=table_id)
+
+    with case(compararison_between_lines, True):
+        filepath = clean_data(
+            file_name=table_id,
+        )
+
+        wait_upload_table = create_table_and_upload_to_gcs(
+            data_path=filepath,
+            dataset_id=dataset_id,
+            table_id=table_id,
+            dump_mode="append",
+            wait=filepath,
+        )
+
+        with case(materialize_after_dump, True):
+            # Trigger DBT flow run
+            current_flow_labels = get_current_flow_labels()
+            materialization_flow = create_flow_run(
+                flow_name=utils_constants.FLOW_EXECUTE_DBT_MODEL_NAME.value,
+                project_name=constants.PREFECT_DEFAULT_PROJECT.value,
+                parameters={
+                    "dataset_id": dataset_id,
+                    "table_id": table_id,
+                    "target": target,
+                    "dbt_alias": dbt_alias,
+                },
+                labels=current_flow_labels,
+                run_name=f"Materialize {dataset_id}.{table_id}",
+            )
+
+            wait_for_materialization = wait_for_flow_run(
+                materialization_flow,
+                stream_states=True,
+                stream_logs=True,
+                raise_final_state=True,
+            )
+            wait_for_materialization.max_retries = (
+                dump_db_constants.WAIT_FOR_MATERIALIZATION_RETRY_ATTEMPTS.value
+            )
+            wait_for_materialization.retry_delay = timedelta(
+                seconds=dump_db_constants.WAIT_FOR_MATERIALIZATION_RETRY_INTERVAL.value
+            )
+
+            with case(update_metadata, True):
+                update_django_metadata(
+                    dataset_id=dataset_id,
+                    table_id=table_id,
+                    date_column_name={"year": "ano", "month": "mes"},
+                    date_format="%Y-%m",
+                    coverage_type="all_free",
+                    prefect_mode=target,
+                    bq_project="basedosdados",
+                    upstream_tasks=[wait_for_materialization],
+                )
+
+flow_isp.storage = GCS(constants.GCS_FLOWS_BUCKET.value)
+flow_isp.run_config = KubernetesRun(image=constants.DOCKER_IMAGE.value)
